requirements.txt
flask==3.0.0
openai==1.10.0
twilio==8.11.0
python-dotenv==1.0.0
requests==2.31.0        
pydub==0.25.1     
psycopg2-binary
